Add safe logout
Logging out via a GET is a poor security practice as it's a stateful
operation.

See OWASP here: https://cheatsheetseries.owasp.org/cheatsheets/Cross-Site_Request_Forgery_Prevention_Cheat_Sheet.html

> Do not use GET requests for state changing operations.
> - If for any reason you do it, protect those resources against CSRF

The easiest way to avoid this is to change the Logout from a loader that
GETs a state changing operation into a form that POSTs.